repos/HardwareAuditUSB/requirements.txt
customtkinter
psutil
py-cpuinfo
GPUtil
packaging
